+++ packages/core/src/generators/stencil/helpers/index.ts
@@ -51,7 +51,7 @@ export const getPropsAsCode = ({
         // Stencil adds "on" to every `@Event` so we need to remove "on" from event props
         // https://stenciljs.com/docs/events#using-events-in-jsx
         const eventType = hasTyping
-          ? `EventEmitter<ReturnType<Required<${propsTypeRef}>["${item}"]>>`
+          ? `EventEmitter<Parameters<Required<${propsTypeRef}>["${item}"]>[number]> | void`
           : 'any';
 
         return `@Event() ${getEventNameWithoutOn(item)}: ${eventType}${defaultPropString}`;